requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
comm==0.1.2
confection==0.0.3
contourpy==1.0.6
cycler==0.11.0
cymem==2.0.7
datasets==2.7.1
debugpy==1.6.4
decorator==5.1.1
dill==0.3.6
dm-tree==0.1.8
docker-pycreds==0.4.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.2
flatbuffers==22.12.6
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
huggingface-hub==0.11.1
idna==3.4
ipykernel==6.19.2
ipython==8.7.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.0
kaleido==0.2.1
keras==2.11.0
kiwisolver==1.4.4
langcodes==3.3.0
libclang==14.0.6
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
multidict==6.0.3
multiprocess==0.70.14
murmurhash==1.0.9
nbformat==5.7.0
nest-asyncio==1.5.6
nltk==3.7
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
parso==0.8.3
pathtools==0.1.2
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
plotex==0.0.8
plotly==5.11.0
preshed==3.0.8
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
Pygments==2.13.0
pynndescent==0.5.8
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytorch_revgrad==0.2.0
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rouge==1.0.1
rouge-score==0.1.2
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.11.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smart-open==6.2.0
smmap==5.0.0
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.4
srsly==2.4.5
stack-data==0.6.2
tenacity==8.1.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
thinc==8.1.5
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.7.0
transformers==4.25.1
typer==0.7.0
typing_extensions==4.4.0
umap-learn==0.5.3
urllib3==1.26.13
wandb==0.13.6
wasabi==0.10.1
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
xxhash==3.1.0
yarl==1.8.2